feb 14 commit for testing

+++ app/src/main/java/nerdvana/com/pointofsales/api_responses/FetchProductsResponse.java
@@ -43,7 +43,7 @@ public class FetchProductsResponse {
         private String productTags;
         @SerializedName("qty")
         @Expose
-        private int qty;
+        private Double qty;
         @SerializedName("tax_rate")
         @Expose
         private double taxRate;
@@ -146,11 +146,11 @@ public class FetchProductsResponse {
             this.productTags = productTags;
         }
 
-        public int getQty() {
+        public Double getQty() {
             return qty;
         }
 
-        public void setQty(int qty) {
+        public void setQty(Double qty) {
             this.qty = qty;
         }
 
@@ -292,7 +292,7 @@ public class FetchProductsResponse {
         private int currencyId;
         @SerializedName("qty")
         @Expose
-        private int qty;
+        private Double qty;
         @SerializedName("price")
         @Expose
         private double price;
@@ -344,11 +344,11 @@ public class FetchProductsResponse {
             this.currencyId = currencyId;
         }
 
-        public int getQty() {
+        public Double getQty() {
             return qty;
         }
 
-        public void setQty(int qty) {
+        public void setQty(Double qty) {
             this.qty = qty;
         }
 
@@ -518,7 +518,7 @@ public class FetchProductsResponse {
 
     public static class BranchGroup {
 
-        public BranchGroup(int selectedQtyInBranch, int id, int coreId, int productId, String groupName, int currencyId, int qty, String createdBy, String createdAt, String updatedAt, String deletedAt) {
+        public BranchGroup(int selectedQtyInBranch, int id, int coreId, int productId, String groupName, int currencyId, Double qty, String createdBy, String createdAt, String updatedAt, String deletedAt) {
             this.selectedQtyInBranch = selectedQtyInBranch;
             this.id = id;
             this.coreId = coreId;
@@ -552,7 +552,7 @@ public class FetchProductsResponse {
         private int currencyId;
         @SerializedName("qty")
         @Expose
-        private int qty;
+        private Double qty;
         @SerializedName("created_by")
         @Expose
         private String createdBy;
@@ -617,11 +617,11 @@ public class FetchProductsResponse {
             this.currencyId = currencyId;
         }
 
-        public int getQty() {
+        public Double getQty() {
             return qty;
         }
 
-        public void setQty(int qty) {
+        public void setQty(Double qty) {
             this.qty = qty;
         }
 
@@ -693,7 +693,7 @@ public class FetchProductsResponse {
         private String productTags;
         @SerializedName("qty")
         @Expose
-        private Integer qty;
+        private Double qty;
         @SerializedName("min")
         @Expose
         private Integer min;
@@ -807,11 +807,11 @@ public class FetchProductsResponse {
             this.productTags = productTags;
         }
 
-        public Integer getQty() {
+        public Double getQty() {
             return qty;
         }
 
-        public void setQty(Integer qty) {
+        public void setQty(Double qty) {
             this.qty = qty;
         }
